Cesar

+++ src/ui/WindowController.java
@@ -1,5 +1,12 @@
 package ui;
 
+/*
+ * Song Library
+ * Cesar Rojas Herrera
+ * Abhishek Panwala
+ * 
+ * 
+ * */
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
@@ -8,7 +15,6 @@ import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Collections;
 
 import be.Song;
 import javafx.collections.FXCollections;
@@ -20,6 +26,8 @@ import javafx.scene.control.ListCell;
 import javafx.scene.control.ListView;
 import javafx.scene.control.TextField;
 import javafx.scene.control.Alert.AlertType;
+import javafx.scene.control.ButtonBar;
+import javafx.scene.control.ButtonType;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.text.Text;
 import javafx.scene.text.TextAlignment;
@@ -27,7 +35,8 @@ import javafx.util.Callback;
 
 public class WindowController {
 	
-	String songFile = "src/Songs.txt";
+	String songFile = "src/data/Songs.txt";
+	boolean confirmB;
 
 	@FXML
     private Text songText;
@@ -167,12 +176,20 @@ public class WindowController {
      * */
     @FXML
     void AddOnAction(ActionEvent event) {
+    	/*Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+    	alert.setTitle("Current song is added");
+    	alert.setContentText("Song is getting added to the list");*/
+    	
+    	displayAlertConfirm("Add Song", "Song being added");
+    	if(confirmB) {
+    		//System.out.println("Yes");
+    	
     	Song song;
     	
     	String Sname = name.getText().trim();
     	String Sartist = artist.getText().trim();
     	String Salbum = album.getText().trim();
-    	String Syear = album.getText().trim();
+    	String Syear = year.getText().trim();
     
     		song = new Song(Sname, Sartist, Salbum, Syear);
     		
@@ -181,7 +198,7 @@ public class WindowController {
     		}
     		try {
     			if(sortInsert(song)) {
-    				displayAlertConfirm("Add Song","Song being added: " + Sname);
+    				//displayAlertConfirm("Add Song","Song being added: " + Sname);
     				BufferedWriter writer = new BufferedWriter(new FileWriter(songFile));
     				
     				for(int i = 0; i < songobsList.size(); i++) {
@@ -197,6 +214,12 @@ public class WindowController {
     		catch(Exception e) {
     			System.out.println(e.toString());
     		}
+    		confirmB = false;
+    	}
+    	else {
+    		confirmB = false;
+    		return;
+    	}
     }
     
     private boolean sortInsert(Song song) {
@@ -244,19 +267,29 @@ public class WindowController {
       * */
     @FXML
     void deleteOnAction(ActionEvent event) {
-    	Song song = songList.getSelectionModel().getSelectedItem();
-    	try {
-    		songobsList.remove(song);
-    		BufferedWriter writer = new BufferedWriter(new FileWriter(songFile));
-			
-			for(int i = 0; i < songobsList.size(); i++) {
-				writer.write(songobsList.get(i).toString());
-			}
-			writer.close();
-			//songList.getSelectionModel().select(songList.getSelectionModel().getSelectedIndex());
+    	//ALERT HERE
+    	displayAlertConfirm("Delete song", "Song being deleted");
+    	if(confirmB) {
+    	
+	    	Song song = songList.getSelectionModel().getSelectedItem();
+	    	try {
+	    		songobsList.remove(song);
+	    		BufferedWriter writer = new BufferedWriter(new FileWriter(songFile));
+				
+				for(int i = 0; i < songobsList.size(); i++) {
+					writer.write(songobsList.get(i).toString());
+				}
+				writer.close();
+				//songList.getSelectionModel().select(songList.getSelectionModel().getSelectedIndex());
+	    	}
+	    	catch(Exception e) {
+	    		System.out.println(e.toString());
+	    	}
+	    	confirmB = false;
     	}
-    	catch(Exception e) {
-    		System.out.println(e.toString());
+    	else {
+    		confirmB = false;
+    		return;
     	}
     }
     
@@ -267,48 +300,59 @@ public class WindowController {
      * */
     @FXML
     void editOnAction(ActionEvent event) {
-    	if(!inputCheck()) {
+    	//ALERT HERE
+    	displayAlertConfirm("Edited song","Song being edited");
+    	
+    	if(confirmB) {
+    	
+	    	if(!inputCheck()) {
+	    		return;
+	    	}
+	    	
+	    	Song song = songList.getSelectionModel().getSelectedItem();
+	    	
+	    	try {
+	    		Song temp = new Song(name.getText().trim(), artist.getText().trim(),
+	    				album.getText().trim(), year.getText().trim());
+	    		if(song.getName().equals(name.getText()) && song.getArtist().equals(artist.getText())) {
+	    			song.setAlbum(album.getText());
+	    			song.setYear(year.getText());
+	    			
+	    			BufferedWriter writer = new BufferedWriter(new FileWriter(songFile));
+	    			
+	    			for(int i = 0; i < songobsList.size(); i++) {
+	    				writer.write(songobsList.get(i).toString());
+	    			}
+	    			writer.close();
+	    			
+	    			clearText();
+	    			//displayAlert("Info", "Song Edited");
+	    		}
+	    		else if(sortInsert(temp)) {
+	    			songobsList.remove(song);
+	    			BufferedWriter writer = new BufferedWriter(new FileWriter(songFile));
+					
+					for(int i = 0; i < songobsList.size(); i++) {
+						writer.write(songobsList.get(i).toString());
+					}
+					writer.close();
+					
+					clearText();
+					//displayAlert("Info","Song Edited");
+	    		}
+	    		else {
+	    			displayAlert("Error","Both song and artist mus be unique");
+	    		}
+	    		
+	    	}
+	    	catch(Exception e) {
+	    		System.out.println(e.toString());
+	    	}
+	    	confirmB = false;
+    	}
+    	else {
+    		confirmB = false;
     		return;
-    	}
-    	
-    	Song song = songList.getSelectionModel().getSelectedItem();
-    	
-    	try {
-    		Song temp = new Song(name.getText().trim(), artist.getText().trim(),
-    				album.getText().trim(), year.getText().trim());
-    		if(song.getName().equals(name.getText()) && song.getArtist().equals(artist.getText())) {
-    			song.setAlbum(album.getText());
-    			song.setYear(year.getText());
-    			
-    			BufferedWriter writer = new BufferedWriter(new FileWriter(songFile));
-    			
-    			for(int i = 0; i < songobsList.size(); i++) {
-    				writer.write(songobsList.get(i).toString());
-    			}
-    			writer.close();
-    			
-    			clearText();
-    			displayAlert("Info", "Song Edited");
-    		}
-    		else if(sortInsert(temp)) {
-    			songobsList.remove(song);
-    			BufferedWriter writer = new BufferedWriter(new FileWriter(songFile));
-				
-				for(int i = 0; i < songobsList.size(); i++) {
-					writer.write(songobsList.get(i).toString());
-				}
-				writer.close();
-				
-				clearText();
-				displayAlert("Info","Song Edited");
-    		}
-    		else {
-    			displayAlert("Error","Both song and artist mus be unique");
-    		}
-    		
-    	}
-    	catch(Exception e) {
-    		System.out.println(e.toString());
     	}
     	
     }
@@ -317,13 +361,13 @@ public class WindowController {
     /*
      * Alert 
      * */
-    private void displayAlert(String string) {
+    /*private void displayAlert(String string) {
     	Alert alert = new Alert(AlertType.WARNING);
     	alert.setTitle("Error!");
     	alert.setHeaderText("Error!");
     	alert.setContentText(string);
     	alert.showAndWait();
-    }
+    }*/
     
     private boolean inputCheck() {
     	
@@ -331,13 +375,17 @@ public class WindowController {
     		displayAlert("Error!", "Not name or artist");
         	return false;
     	}
-    	else {
+    	else{
+    		if(Integer.valueOf(year.getText()) < 0) { 
+    			displayAlert("Error", "Year value has to be positive");
+    			return false;
+    		}
     		return true;
     	}
     }
     
     private void displayAlert(String header, String content) {
-    	Alert alert = new Alert(AlertType.INFORMATION);
+    	Alert alert = new Alert(AlertType.ERROR);
     	alert.setTitle("Alert");
     	alert.setHeaderText(header);
     	alert.setContentText(content);
@@ -348,11 +396,27 @@ public class WindowController {
      * Alert with warning?
      * */
     private void displayAlertConfirm(String header, String content) {
-    	Alert alert = new Alert(AlertType.WARNING);
+    	Alert alert = new Alert(AlertType.CONFIRMATION);
     	alert.setTitle("Alert");
     	alert.setHeaderText(header);
     	alert.setContentText(content);
-    	alert.showAndWait();
+    	
+    	ButtonType okB = new ButtonType("Yes", ButtonBar.ButtonData.YES);
+    	//ButtonType noB = new ButtonType("No", ButtonBar.ButtonData.NO);
+    	ButtonType cancelB = new ButtonType ("Cancel", ButtonBar.ButtonData.CANCEL_CLOSE);
+    	alert.getButtonTypes().setAll(okB, cancelB);
+    	alert.showAndWait().ifPresent(type ->{
+    		if(type == okB) {
+    			//System.out.println("Ok");
+    			confirmB = true;
+    		}
+    		else if(type == cancelB) {
+    			//System.out.println("No");
+    			confirmB = false;
+    		}
+    	});
+    	
+    	
     }
     
     private void setText() {
